Update LinuxSoftwareInstaller.java
Fixed uninstallPackage to have correct behavior. A package should be evaluated if it is needed as a dependency for any other installedPackage, and if not, it is added to a list packagesToUninstall that will be removed all at once. If there is an issue this list can just be cleared, instead of uninstalling/re-installing to a restore point.